Fix more Windows 32 and 64 bit compiler warnings introduced by revision 7874 WL6968  InnoDB GIS: R-tree index support. WL6609: InnoDB GIS: Support Predicate Locking for GIS index WL6745: InnoDB GIS: Support DML for InnoDB GIS index

+++ storage/innobase/trx/trx0rec.cc
@@ -549,8 +549,10 @@ trx_undo_get_mbr_from_ext(
 			mbr[i * 2 + 1] = -DBL_MAX;
 		}
 	} else {
-		rtree_mbr_from_wkb(dptr + GEO_DATA_HEADER_SIZE,
-				   dlen - GEO_DATA_HEADER_SIZE, SPDIMS, mbr);
+		rtree_mbr_from_wkb(
+			dptr + GEO_DATA_HEADER_SIZE,
+			static_cast<uint>(dlen) - GEO_DATA_HEADER_SIZE,
+			SPDIMS, mbr);
 	}
 
 	mem_heap_free(heap);
